requirements.txt
click==8.1.7
evidently==0.4.13
hydra-core==1.3.2
Markdown==3.5.1
matplotlib>=3.7.0
monai==1.3.0
monai_weekly==1.4.dev2401
numpy==1.26.3
pandas>=2.0.0
PyYAML==6.0.1
torch==2.1.2
torchvision==0.16.2
wandb==0.16.2
